requirements.txt
# Construction Planning Assistant — Dependencies
# Install with: pip install -r requirements.txt

# API server
fastapi==0.135.3
uvicorn==0.44.0

# LLM provider
groq==1.1.2

# Data validation
pydantic==2.12.5

# Environment variables
python-dotenv==1.2.2

# Google ADK (used in agents/ — optional if running Groq-only mode)
google-adk==1.29.0
